requirements.txt
streamlit
groq
PyPDF2
Pillow
python-dotenv